7.1.4

+++ com/baidu/tieba/im/db/e.java
@@ -13,7 +13,7 @@ import java.util.LinkedList;
 /* loaded from: classes.dex */
 public class e extends SQLiteOpenHelper {
     public e(Context context, String str) {
-        super(context, str, (SQLiteDatabase.CursorFactory) null, 10);
+        super(context, str, (SQLiteDatabase.CursorFactory) null, 11);
     }
 
     private void e(SQLiteDatabase sQLiteDatabase, String str) {
@@ -37,7 +37,7 @@ public class e extends SQLiteOpenHelper {
 
     private void s(SQLiteDatabase sQLiteDatabase) {
         if (sQLiteDatabase != null) {
-            e(sQLiteDatabase, "CREATE TABLE IF NOT EXISTS tb_message_center(gid TEXT NOT NULL, group_name TEXT, group_head TEXT, custom_group_type int, group_ext TEXT, unread_count int, last_msgId TEXT, last_user_name TEXT, last_content_time long, send_status int, last_content TEXT, is_friend int, pull_msgid TEXT,is_hidden int,is_delete int, sent_mid long, read_mid long, user_type int default 1);");
+            e(sQLiteDatabase, "CREATE TABLE IF NOT EXISTS tb_message_center(gid TEXT NOT NULL, group_name TEXT, group_head TEXT, custom_group_type int, group_ext TEXT, unread_count int, last_msgId TEXT, last_user_name TEXT, last_content_time long, send_status int, last_content TEXT, is_friend int, pull_msgid TEXT,is_hidden int,is_delete int, sent_mid long, read_mid long, sid long, task_id TEXT, ext1 TEXT, ext2 TEXT, user_type int default 1);");
         }
     }
 
@@ -76,6 +76,7 @@ public class e extends SQLiteOpenHelper {
                     C(sQLiteDatabase);
                     D(sQLiteDatabase);
                     E(sQLiteDatabase);
+                    F(sQLiteDatabase);
                     return;
                 } catch (Exception e) {
                     TiebaStatic.printDBExceptionLog(e, "ImDatabaseHelper.onUpgrade", new Object[0]);
@@ -92,6 +93,7 @@ public class e extends SQLiteOpenHelper {
                 C(sQLiteDatabase);
                 D(sQLiteDatabase);
                 E(sQLiteDatabase);
+                F(sQLiteDatabase);
                 return;
             case 3:
                 w(sQLiteDatabase);
@@ -101,6 +103,7 @@ public class e extends SQLiteOpenHelper {
                 C(sQLiteDatabase);
                 D(sQLiteDatabase);
                 E(sQLiteDatabase);
+                F(sQLiteDatabase);
                 return;
             case 4:
                 x(sQLiteDatabase);
@@ -109,6 +112,7 @@ public class e extends SQLiteOpenHelper {
                 C(sQLiteDatabase);
                 D(sQLiteDatabase);
                 E(sQLiteDatabase);
+                F(sQLiteDatabase);
                 return;
             case 5:
                 y(sQLiteDatabase);
@@ -116,24 +120,32 @@ public class e extends SQLiteOpenHelper {
                 C(sQLiteDatabase);
                 D(sQLiteDatabase);
                 E(sQLiteDatabase);
+                F(sQLiteDatabase);
                 return;
             case 6:
                 B(sQLiteDatabase);
                 C(sQLiteDatabase);
                 D(sQLiteDatabase);
                 E(sQLiteDatabase);
+                F(sQLiteDatabase);
                 return;
             case 7:
                 C(sQLiteDatabase);
                 D(sQLiteDatabase);
                 E(sQLiteDatabase);
+                F(sQLiteDatabase);
                 return;
             case 8:
                 D(sQLiteDatabase);
                 E(sQLiteDatabase);
+                F(sQLiteDatabase);
                 return;
             case 9:
                 E(sQLiteDatabase);
+                F(sQLiteDatabase);
+                return;
+            case 10:
+                F(sQLiteDatabase);
                 return;
             default:
                 return;
@@ -196,7 +208,7 @@ public class e extends SQLiteOpenHelper {
                 Iterator<String> it = z.iterator();
                 while (it.hasNext()) {
                     String next = it.next();
-                    if (!TextUtils.isEmpty(next) && (next.startsWith(l.bMo) || next.startsWith(k.bMo))) {
+                    if (!TextUtils.isEmpty(next) && (next.startsWith(l.bPU) || next.startsWith(k.bPU))) {
                         try {
                             sQLiteDatabase.execSQL("ALTER TABLE " + next + " ADD is_friend int default 1;");
                         } catch (Exception e) {
@@ -299,12 +311,12 @@ public class e extends SQLiteOpenHelper {
                 sQLiteDatabase.setTransactionSuccessful();
                 try {
                     sQLiteDatabase.beginTransaction();
-                    f.bMu.put(String.valueOf(TbadkCoreApplication.getCurrentAccount()) + ".db", sQLiteDatabase);
-                    j.Xp();
+                    f.bQa.put(String.valueOf(TbadkCoreApplication.getCurrentAccount()) + ".db", sQLiteDatabase);
+                    j.Yv();
                     sQLiteDatabase.setTransactionSuccessful();
                 } finally {
                     sQLiteDatabase.endTransaction();
-                    f.bMu.remove(String.valueOf(TbadkCoreApplication.getCurrentAccount()) + ".db");
+                    f.bQa.remove(String.valueOf(TbadkCoreApplication.getCurrentAccount()) + ".db");
                 }
             } finally {
                 sQLiteDatabase.endTransaction();
@@ -356,6 +368,26 @@ public class e extends SQLiteOpenHelper {
                     sQLiteDatabase.execSQL("CREATE TABLE IF NOT EXISTS tb_new_friends(_id INTEGER primary key autoincrement, uid LONG NOT NULL, uname TEXT, ustatus INTEGER, uportrait TEXT, ucontent TEXT,isread INTEGER);");
                     sQLiteDatabase.execSQL("INSERT INTO tb_new_friends SELECT _id,uid,uname,ustatus,uportrait,ucontent,isread FROM tb_new_friends_tmp WHERE uid > 0;");
                     sQLiteDatabase.execSQL("DROP TABLE tb_new_friends_tmp;");
+                } catch (Exception e) {
+                    TiebaStatic.printDBExceptionLog(e, "ImDatabaseHelper.groupMsg9to10", new Object[0]);
+                    sQLiteDatabase.execSQL("DROP TABLE IF EXISTS tb_new_friends");
+                }
+                sQLiteDatabase.setTransactionSuccessful();
+            } finally {
+                sQLiteDatabase.endTransaction();
+            }
+        }
+    }
+
+    private void F(SQLiteDatabase sQLiteDatabase) {
+        if (sQLiteDatabase != null) {
+            try {
+                sQLiteDatabase.beginTransaction();
+                try {
+                    sQLiteDatabase.execSQL("ALTER TABLE tb_message_center ADD sid long default 0;");
+                    sQLiteDatabase.execSQL("ALTER TABLE tb_message_center ADD task_id TEXT;");
+                    sQLiteDatabase.execSQL("ALTER TABLE tb_message_center ADD ext1 TEXT;");
+                    sQLiteDatabase.execSQL("ALTER TABLE tb_message_center ADD ext2 TEXT;");
                 } catch (Exception e) {
                     TiebaStatic.printDBExceptionLog(e, "ImDatabaseHelper.groupMsg9to10", new Object[0]);
                     sQLiteDatabase.execSQL("DROP TABLE IF EXISTS tb_new_friends");
